Support reconnect attempt on request failure (#49)
Attempt to reconnect to the protocol adapter once after a failure to complete a request (for any reason) after retrying the set number of times. If reconnected, attempt the request again.
Raise full ScreenLogicError if reconnect or second full request attempt fails. This supports Home Assistant registering the failure and handling it accordingly.